backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pandas==2.2.2
numpy==1.26.4
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
joblib==1.4.2
scikit-learn==1.5.0
scipy==1.13.1
pillow==10.3.0
oracledb==2.4.1
